backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.2
pydantic[email]==2.9.2
bleach==6.1.0
